+++ fcxref/remove/__init__.py
@@ -1,0 +1,3 @@
+from .remove import make_remove
+
+__all__ = ['make_remove']