+++ tests/find_test.py
@@ -35,33 +35,33 @@ class FindTest(unittest.TestCase):
                                    'MainDocument#Spreadsheet.Value',
                                    xpath0))
 
-        xpath1 = "ObjectData/Object[@name='Spreadsheet']/Properties/Property[@name='cells']/Cells/Cell[@address='A1']"
+        xpath1 = "ObjectData/Object[@name='Box']/Properties/Property[@name='ExpressionEngine']/ExpressionEngine/Expression[@path='Length']"
         self.assertEqual(references[1],
+                         Reference('ExampleDocument.FCStd',
+                                   'Box',
+                                   'ExpressionEngine',
+                                   'expression',
+                                   'Length',
+                                   'MainDocument#Spreadsheet.Value',
+                                   xpath1))
+
+        xpath2 = "ObjectData/Object[@name='Spreadsheet']/Properties/Property[@name='cells']/Cells/Cell[@address='A1']"
+        self.assertEqual(references[2],
                          Reference('ExampleDocument.FCStd',
                                    'Spreadsheet',
                                    'cells',
                                    'content',
                                    'A1',
                                    'Value',
-                                   xpath1))
+                                   xpath2))
 
-        xpath2 = "ObjectData/Object[@name='Spreadsheet']/Properties/Property[@name='cells']/Cells/Cell[@address='B1']"
-        self.assertEqual(references[2],
+        xpath3 = "ObjectData/Object[@name='Spreadsheet']/Properties/Property[@name='cells']/Cells/Cell[@address='B1']"
+        self.assertEqual(references[3],
                          Reference('ExampleDocument.FCStd',
                                    'Spreadsheet',
                                    'cells',
                                    'alias',
                                    'B1',
-                                   'Value',
-                                   xpath2))
-
-        xpath3 = "ObjectData/Object[@name='Box']/Properties/Property[@name='ExpressionEngine']/ExpressionEngine/Expression[@path='Length']"
-        self.assertEqual(references[3],
-                         Reference('ExampleDocument.FCStd',
-                                   'Box',
-                                   'ExpressionEngine',
-                                   'expression',
-                                   'Length',
                                    'Value',
                                    xpath3))